+++ src/main/java/controllers/admin/EmployeeManagerController.java
@@ -30,7 +30,7 @@ public class EmployeeManagerController extends ManagerController {
     @Override
     public void actionEdit() {
         try {
-            int selectedId = view.getSelectedId();
+            long selectedId = view.getSelectedId();
             if (selectedId < 0) {
                 throw new Exception("Chọn nhân viên cần edit");
             }
@@ -48,7 +48,7 @@ public class EmployeeManagerController extends ManagerController {
 
     @Override
     public void actionDelete() {
-        int selectedIds[] = view.getSelectedIds();
+        Long selectedIds[] = view.getSelectedIds();
         try {
             if (JOptionPane.showConfirmDialog(null, "Xác nhận xóa hàng loạt?", "Xóa nhân viên", ERROR_MESSAGE) != YES_OPTION) {
                 return;
