+++ src/main/java/controllers/admin/UserManagerController.java
@@ -11,11 +11,7 @@ import models.User;
 import utils.UserRole;
 import views.popup.EmployeePopupView;
 
-/**
- * createAt Dec 15, 2020
- *
- * @author Đỗ Tuấn Anh <daclip26@gmail.com>
- */
+
 public class UserManagerController extends ManagerController {
 
     UserDao userDao = new UserDao();
@@ -34,7 +30,7 @@ public class UserManagerController extends ManagerController {
     @Override
     public void actionEdit() {
         try {
-            int selectedId = view.getSelectedId();
+            long selectedId = view.getSelectedId();
             if (selectedId < 0) {
                 throw new Exception("Chọn nhân viên cần edit");
             }
@@ -58,7 +54,7 @@ public class UserManagerController extends ManagerController {
 
     @Override
     public void actionDelete() {
-        int selectedIds[] = view.getSelectedIds();
+        Long selectedIds[] = view.getSelectedIds();
         try {
             if (JOptionPane.showConfirmDialog(null, "Xác nhận xóa hàng loạt?", "Xóa nhân viên", ERROR_MESSAGE) != YES_OPTION) {
                 return;
